\name{coef.fastVAR.GraphicalVARX}
\alias{coef.fastVAR.GraphicalVARX}
\title{GraphicalVARX Coefficients}
\usage{
  coef.fastVAR.GraphicalVARX(GraphicalVARX, ...)
}
\arguments{
  \item{VARX}{an object of class fastVAR.VARX}

  \item{...}{}
}
\value{
  The coefficients for the VARX model
}
\description{
  GraphicalVARX Coefficients
}

