\name{coef.fastVAR.GroupVAR}
\alias{coef.fastVAR.GroupVAR}
\title{GroupVAR Coefficients}
\usage{
  \method{coef}{fastVAR.GroupVAR} (GroupVAR, ...)
}
\arguments{
  \item{GroupVAR}{an object of class fastVAR.GroupVAR}

  \item{...}{extra parameters to pass into the coefficients
  method for objects of type fastVAR.GroupVAR}
}
\value{
  The coefficients for the VAR model
}
\description{
  GroupVAR Coefficients
}

