\name{coef.fastVAR.GroupVARX}
\alias{coef.fastVAR.GroupVARX}
\title{GroupVARX Coefficients}
\usage{
  \method{coef}{fastVAR.GroupVARX} (GroupVARX, ...)
}
\arguments{
  \item{GroupVARX}{an object of class fastVAR.GroupVARX}

  \item{...}{extra parameters to pass into the coefficients
  method for objects of type fastVAR.GroupVARX}
}
\value{
  The coefficients for the GroupVARX model
}
\description{
  GroupVARX Coefficients
}

