\name{deseason}
\alias{deseason}
\title{Deseason}
\usage{
  deseason(mts, frequency = NA, auto = F)
}
\arguments{
  \item{mts}{either a vector representing 1 time series or
  a a data frame or matrix representing multiple time
  series}
}
\description{
  Deseasonalize a time series so that models can target the
  unexplainable components.
}

